Update DrivewayViewController.swift
Adding loader along with db accessor to retrieve the state of the segment control to display the status of the driveway when user opens view. This is done by adding a loader UIView along with a new method that retrieves the status of the segment control store in firebase. I retrieve the value and update in the viewWillAppear(...) Method.